+++ CHANGELOG.md
@@ -21,6 +21,7 @@ _This release is scheduled to be released on 2024-10-01._
 - [weather] Updated `apiVersion` default from 2.5 to 3.0 (#3424)
 - [core] Updated dependencies including stylistic-eslint
 - [core] Allow custom module positions by setting `allowCustomModulePositions` in `config.js` (fixes #3504, related to https://github.com/MagicMirrorOrg/MagicMirror/pull/3445)
+- [core] Updated SocketIO catch all to new API
 
 ### Fixed